+++ src/main/java/com/sg/banco/domain/CheckingAccount.java
@@ -5,6 +5,7 @@ import lombok.*;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.PrimaryKeyJoinColumn;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -17,13 +18,13 @@ import java.time.LocalDate;
 public class CheckingAccount extends Account {
 
     @Column(name = "overdraft_limit")//limite
-    private Double overdraftLimit;
+    private BigDecimal overdraftLimit;
 
     @Column(name = "interest_rate")//taxa de juros
-    private Double interestRate;
+    private BigDecimal interestRate;
 
     @Column//juros
-    private Double interest;
+    private BigDecimal interest;
 
     @Column(name = "interest_day")//dia que ultrapassou o limite
     private LocalDate interestDay;
